third_party/blink/web_tests/platform/generic/fast/dom/DOMException/stack-trace-expected.txt

PASS "stack" in e is true
PASS typeof e.stack is "string"
PASS e.stack.contains("innerFunction") is true
PASS e.stack.contains("outerFunction") is true
PASS Object.prototype.toString.call(e) is "[object DOMException]"
PASS e.toString() is "HierarchyRequestError: Failed to execute 'appendChild' on 'Node': Nodes of type '#document' may not be inserted inside nodes of type '#document'."
PASS e.name is "HierarchyRequestError"
PASS e.message is "Failed to execute 'appendChild' on 'Node': Nodes of type '#document' may not be inserted inside nodes of type '#document'."
PASS e.code is 3
PASS e.stack === 42 is true
PASS successfullyParsed is true

TEST COMPLETE
